// JavaScript/src/dotnet-types.d.ts
interface LoadingResource {
    name: string;
    url: string;
    response: Promise<Response>;
}
interface AssetEntry {
    /**
     * the name of the asset, including extension.
     */
    name: string;
    /**
     * determines how the asset will be handled once loaded
     */
    behavior: AssetBehaviors;
    /**
     * this should be absolute url to the asset
     */
    resolvedUrl?: string;
    /**
     * the integrity hash of the asset (if any)
     */
    hash?: string | null | "";
    /**
     * If specified, overrides the path of the asset in the virtual filesystem and similar data structures once downloaded.
     */
    virtualPath?: string;
    /**
     * Culture code
     */
    culture?: string;
    /**
     * If true, an attempt will be made to load the asset from each location in MonoConfig.remoteSources.
     */
    loadRemote?: boolean;
    /**
     * If true, the runtime startup would not fail if the asset download was not successful.
     */
    isOptional?: boolean;
    /**
     * If provided, runtime doesn't have to fetch the data.
     * Runtime would set the buffer to null after instantiation to free the memory.
     */
    buffer?: ArrayBuffer | Promise<ArrayBuffer>;
    /**
     * If provided, runtime doesn't have to import it's JavaScript modules.
     * This will not work for multi-threaded runtime.
     */
    moduleExports?: any | Promise<any>;
    /**
     * It's metadata + fetch-like Promise<Response>
     * If provided, the runtime doesn't have to initiate the download. It would just await the response.
     */
    pendingDownload?: LoadingResource;
}
type SingleAssetBehaviors =
/**
 * The binary of the dotnet runtime.
 */
    "dotnetwasm"
    /**
     * The javascript module for loader.
     */
    | "js-module-dotnet"
    /**
     * The javascript module for threads.
     */
    | "js-module-threads"
    /**
     * The javascript module for runtime.
     */
    | "js-module-runtime"
    /**
     * The javascript module for emscripten.
     */
    | "js-module-native"
    /**
     * Typically blazor.boot.json
     */
    | "manifest";
type AssetBehaviors = SingleAssetBehaviors |
    /**
     * Load asset as a managed resource assembly.
     */
    "resource"
    /**
     * Load asset as a managed assembly.
     */
    | "assembly"
    /**
     * Load asset as a managed debugging information.
     */
    | "pdb"
    /**
     * Store asset into the native heap.
     */
    | "heap"
    /**
     * Load asset as an ICU data archive.
     */
    | "icu"
    /**
     * Load asset into the virtual filesystem (for fopen, File.Open, etc).
     */
    | "vfs"
    /**
     * The javascript module that came from nuget package .
     */
    | "js-module-library-initializer"
    /**
     * The javascript module for threads.
     */
    | "symbols";